Apply a pattern to BigDecimals  operations and some tests

+++ score-module-application/src/main/java/ifood/score/domain/utils/ScoreBuilder.java
@@ -31,7 +31,7 @@ public class ScoreBuilder {
         validate();
 
         BigDecimal sum = relevanceItems.stream().map(Relevance.Item::getValue).reduce((v, acc) -> acc = acc.add(v)).get();
-        BigDecimal scoreVal = MathOperations.divide(sum, BigDecimal.valueOf(relevanceItems.size()));
+        BigDecimal scoreVal = MathOperations.divide(sum, relevanceItems.size());
 
         Score score;
         Category category = relevanceItems.get(0).getCategory();
